latest non-async

+++ Source/Podio .NET/Models/Request/BulkDeleteRequest.cs
@@ -10,7 +10,7 @@ namespace PodioAPI.Models.Request
         ///     An explicit list of item ids to be deleted
         /// </summary>
         [JsonProperty(PropertyName = "item_ids", NullValueHandling = NullValueHandling.Ignore)]
-        public List<long> ItemIds { get; set; }
+        public List<int> ItemIds { get; set; }
 
         /// <summary>
         ///     The id of the saved view whose items are to be deleted. Ignored if "item_ids" is given.
